Optimize integer comparisons between fixnum and bignum
Bignums should never be in the NAT_MIN_FIXNUM..NAT_MAX_FIXNUM range, so
we can optimize some comparisons between the different backing types.

+++ include/natalie/bigint.hpp
@@ -206,7 +206,7 @@ struct BigInt {
         return strtol(buf, nullptr, 10);
     }
 
-    long to_long_long() const {
+    long long to_long_long() const {
         char buf[32];
         bigint_write(buf, 32, data);
         return strtoll(buf, nullptr, 10);
